@@ -46,21 +46,19 @@
     "release.demo": "node --import tsx scripts/src/release-demo.mts"
   },
   "dependencies": {
-    "@angular/animations": "^18.2.1",
-    "@angular/cdk": "^18.2.1",
-    "@angular/common": "^18.2.1",
-    "@angular/compiler": "^18.2.1",
-    "@angular/core": "^18.2.1",
-    "@angular/forms": "^18.2.1",
-    "@angular/material": "^18.2.1",
-    "@angular/platform-browser": "^18.2.1",
-    "@angular/platform-browser-dynamic": "^18.2.1",
-    "@angular/router": "^18.2.1",
+    "@angular/animations": "^19.1.5",
+    "@angular/cdk": "^19.1.3",
+    "@angular/common": "^19.1.5",
+    "@angular/compiler": "^19.1.5",
+    "@angular/core": "^19.1.5",
+    "@angular/forms": "^19.1.5",
+    "@angular/material": "^19.1.3",
+    "@angular/platform-browser": "^19.1.5",
+    "@angular/platform-browser-dynamic": "^19.1.5",
+    "@angular/router": "^19.1.5",
     "core-js": "^3.32.2",
     "file-saver": "^2.0.2",
-    "hammerjs": "2.0.8",
     "html2canvas": "^1.4.1",
-    "jspdf": "^2.5.1",
     "jszip": "^3.10.1",
     "jwt-decode": "^4.0.0",
     "moment": "^2.30.1",
@@ -92,22 +90,20 @@
     "@semantic-release/git": "^10.0.1",
     "@semantic-release/github": "^11.0.0",
     "@sentry/types": "^8.41.0",
-    "@stylistic/eslint-plugin": "^2.11.0",
-    "@trivago/prettier-plugin-sort-imports": "^4.3.0",
+    "@stylistic/eslint-plugin": "^4.0.0",
+    "@trivago/prettier-plugin-sort-imports": "^5.2.0",
     "@types/flexsearch": "^0.7.6",
     "@types/geojson": "^7946.0.10",
-    "@types/hammerjs": "^2.0.36",
     "@types/jasmine": "~5.1.0",
     "@types/lodash-es": "^4.17.0",
-    "@types/node": "^20.10.0",
+    "@types/node": "^22.13.0",
     "@types/tinycolor2": "^1.4.4",
     "angular-cli-ghpages": "^2.0.0",
-    "angular-eslint": "^18.4.0",
+    "angular-eslint": "^19.1.0",
     "conventional-changelog-cli": "^5.0.0",
     "copy-newer": "^2.1.2",
-    "cypress": "^13.16.0",
-    "del": "^8.0.0",
-    "eslint": "^9.15.0",
+    "cypress": "^14.0.0",
+    "eslint": "^9.21.0",
     "execa": "^9.5.0",
     "jasmine-core": "~5.1.0",
     "jasmine-spec-reporter": "~7.0.0",
