@@ -1,6 +1,6 @@
 {
   "name": "@igo2/auth",
-  "version": "18.0.0",
+  "version": "19.0.0-next.5",
   "description": "IGO Library",
   "author": "IGO Community",
   "keywords": [
@@ -55,12 +55,12 @@
     "tslib": "^2.6.0"
   },
   "peerDependencies": {
-    "@angular/cdk": "^18.0.0",
-    "@angular/common": "^18.0.0",
-    "@angular/core": "^18.0.0",
-    "@angular/forms": "^18.0.0",
-    "@angular/material": "^18.0.0",
-    "@angular/router": "^18.0.0",
+    "@angular/cdk": "^19.0.0",
+    "@angular/common": "^19.0.0",
+    "@angular/core": "^19.0.0",
+    "@angular/forms": "^19.0.0",
+    "@angular/material": "^19.0.0",
+    "@angular/router": "^19.0.0",
     "@azure/msal-angular": "^3.0.4",
     "@azure/msal-browser": "^3.0.4",
     "@igo2/common": "*",
