@@ -1,6 +1,6 @@
 {
   "name": "@igo2/common",
-  "version": "18.0.0",
+  "version": "19.0.0-next.5",
   "description": "IGO Library",
   "author": "IGO Community",
   "keywords": [
@@ -142,14 +142,14 @@
     "tslib": "^2.6.0"
   },
   "peerDependencies": {
-    "@angular/common": "^18.0.0",
-    "@angular/core": "^18.0.0",
-    "@angular/material": "^18.0.0",
-    "@angular/platform-browser": "^18.0.0",
+    "@angular/common": "^19.0.0",
+    "@angular/core": "^19.0.0",
+    "@angular/material": "^19.0.0",
+    "@angular/platform-browser": "^19.0.0",
     "@floating-ui/utils": "^0.2.0",
     "@igo2/core": "*",
     "@igo2/utils": "*",
-    "angular-shepherd": "^18.0.0",
+    "angular-shepherd": "^19.0.0",
     "ngx-color": "^9.0.0",
     "scroll-into-view-if-needed": "^3.0.0",
     "tinycolor2": "^1.6.0",
