@@ -1,6 +1,6 @@
 {
   "name": "@igo2/context",
-  "version": "18.0.0",
+  "version": "19.0.0-next.5",
   "description": "IGO Library",
   "author": "IGO Community",
   "keywords": [
@@ -33,12 +33,12 @@
     }
   },
   "peerDependencies": {
-    "@angular/animations": "^18.0.0",
-    "@angular/common": "^18.0.0",
-    "@angular/core": "^18.0.0",
-    "@angular/forms": "^18.0.0",
-    "@angular/material": "^18.0.0",
-    "@angular/platform-browser": "^18.0.0",
+    "@angular/animations": "^19.0.0",
+    "@angular/common": "^19.0.0",
+    "@angular/core": "^19.0.0",
+    "@angular/forms": "^19.0.0",
+    "@angular/material": "^19.0.0",
+    "@angular/platform-browser": "^19.0.0",
     "@igo2/auth": "*",
     "@igo2/common": "*",
     "@igo2/core": "*",
