@@ -1,6 +1,6 @@
 {
   "name": "@igo2/core",
-  "version": "18.0.0",
+  "version": "19.0.0-next.5",
   "description": "IGO Library",
   "author": "IGO Community",
   "keywords": [
@@ -62,9 +62,6 @@
     "./environment": {
       "import": "./environment/src/public_api.ts"
     },
-    "./gesture": {
-      "import": "./gesture/src/public_api.ts"
-    },
     "./language": {
       "import": "./language/src/public_api.ts"
     },
@@ -100,11 +97,11 @@
     "tslib": "^2.6.0"
   },
   "peerDependencies": {
-    "@angular/cdk": "^18.0.0",
-    "@angular/common": "^18.0.0",
-    "@angular/core": "^18.0.0",
-    "@angular/platform-browser": "^18.0.0",
-    "@angular/router": "^18.0.0",
+    "@angular/cdk": "^19.0.0",
+    "@angular/common": "^19.0.0",
+    "@angular/core": "^19.0.0",
+    "@angular/platform-browser": "^19.0.0",
+    "@angular/router": "^19.0.0",
     "@igo2/utils": "*",
     "@sentry/angular": "^8.26.0",
     "ngx-toastr": "^19.0.0",
