@@ -1,6 +1,6 @@
 {
   "name": "@igo2/geo",
-  "version": "18.0.0",
+  "version": "19.0.0-next.5",
   "description": "IGO Library",
   "author": "IGO Community",
   "keywords": [
@@ -47,23 +47,23 @@
     "tslib": "^2.6.0"
   },
   "peerDependencies": {
-    "@angular/animations": "^18.0.0",
-    "@angular/common": "^18.0.0",
-    "@angular/core": "^18.0.0",
-    "@angular/forms": "^18.0.0",
-    "@angular/material": "^18.0.0",
-    "@angular/material-moment-adapter": "^18.0.0",
-    "@angular/platform-browser": "^18.0.0",
+    "@angular/animations": "^19.0.0",
+    "@angular/common": "^19.0.0",
+    "@angular/core": "^19.0.0",
+    "@angular/forms": "^19.0.0",
+    "@angular/material": "^19.0.0",
+    "@angular/material-moment-adapter": "^19.0.0",
+    "@angular/platform-browser": "^19.0.0",
     "@igo2/auth": "*",
     "@igo2/common": "*",
     "@igo2/core": "*",
     "@igo2/utils": "*",
-    "@mat-datetimepicker/core": "~14.0.0",
+    "@mat-datetimepicker/core": "^15.0.1",
     "file-saver": "^2.0.2",
     "flexsearch": "0.7.21",
     "html2canvas": "^1.4.1",
-    "jspdf": "^2.5.1",
-    "jspdf-autotable": "^3.8.0",
+    "jspdf": "^3.0.0",
+    "jspdf-autotable": "^5.0.0",
     "jszip": "^3.10.1",
     "moment": "^2.30.0",
     "ngx-indexed-db": "^19.0.0",
