@@ -1,29 +1,24 @@
 import { AsyncPipe, NgFor, NgIf } from '@angular/common';
 import { Component, Inject, OnDestroy, OnInit, Optional } from '@angular/core';
-import {
-  FormsModule,
-  ReactiveFormsModule,
-  UntypedFormBuilder,
-  UntypedFormGroup,
-  Validators
-} from '@angular/forms';
+import { FormsModule, ReactiveFormsModule, Validators } from '@angular/forms';
 import { MatAutocompleteModule } from '@angular/material/autocomplete';
 import { MatButtonModule } from '@angular/material/button';
 import { MatOptionModule } from '@angular/material/core';
-import {
-  MAT_DIALOG_DATA,
-  MatDialogActions,
-  MatDialogContent,
-  MatDialogRef,
-  MatDialogTitle
-} from '@angular/material/dialog';
+import { MAT_DIALOG_DATA, MatDialogRef } from '@angular/material/dialog';
 import { MatFormFieldModule } from '@angular/material/form-field';
 import { MatInputModule } from '@angular/material/input';
 import { MatListModule } from '@angular/material/list';
 import { MatSelectModule } from '@angular/material/select';
+import { MatTabsModule } from '@angular/material/tabs';
 import { MatTooltipModule } from '@angular/material/tooltip';
 
 import { EntityStore } from '@igo2/common/entity';
+import {
+  Form,
+  FormComponent,
+  FormFieldComponent,
+  FormService
+} from '@igo2/common/form';
 import { ConfigService } from '@igo2/core/config';
 import { LanguageService } from '@igo2/core/language';
 import { IgoLanguageModule } from '@igo2/core/language';
@@ -37,44 +32,37 @@ import { Catalog } from '../shared/catalog.abstract';
   selector: 'igo-add-catalog-dialog',
   templateUrl: './add-catalog-dialog.component.html',
   styleUrls: ['./add-catalog-dialog.component.scss'],
-  standalone: true,
   imports: [
-    MatDialogTitle,
-    MatDialogContent,
     FormsModule,
     ReactiveFormsModule,
     MatFormFieldModule,
     MatInputModule,
     MatAutocompleteModule,
+    MatTabsModule,
     NgFor,
     MatOptionModule,
     MatTooltipModule,
     MatSelectModule,
     MatListModule,
     NgIf,
-    MatDialogActions,
     MatButtonModule,
     AsyncPipe,
-    IgoLanguageModule
+    IgoLanguageModule,
+    FormComponent,
+    FormFieldComponent
   ]
 })
 export class AddCatalogDialogComponent implements OnInit, OnDestroy {
-  public form: UntypedFormGroup;
-  private defaultAddedCatalogType = 'wms';
-  private addedCatalogType$$: Subscription;
-  public predefinedCatalogsList$ = new BehaviorSubject<Catalog[]>([]);
-  typeCapabilities: string[];
-  predefinedCatalogs: Catalog[] = [];
-  store: EntityStore<Catalog>;
-  error = false;
-  addedCatalog: Catalog;
+  predefinedForm$ = new BehaviorSubject<Form>(undefined);
+  customForm$ = new BehaviorSubject<Form>(undefined);
+  data$ = new BehaviorSubject<Catalog>(undefined);
+  customData$ = new BehaviorSubject<Catalog>(undefined);
   emailAddress: string;
   private storeViewAll$$: Subscription;
-
   constructor(
-    private formBuilder: UntypedFormBuilder,
     public languageService: LanguageService,
     private configService: ConfigService,
+    private formService: FormService,
     public dialogRef: MatDialogRef<AddCatalogDialogComponent>,
     @Optional()
     @Inject(MAT_DIALOG_DATA)
@@ -84,66 +72,112 @@ export class AddCatalogDialogComponent implements OnInit, OnDestroy {
       error: boolean;
       addedCatalog: Catalog;
     }
-  ) {
-    this.store = data.store;
-    this.predefinedCatalogs = data.predefinedCatalogs;
-    this.error = data.error;
-    this.addedCatalog = data.addedCatalog;
-    this.form = this.formBuilder.group({
-      id: ['', []],
-      title: ['', []],
-      url: ['', [Validators.required]],
-      type: [this.defaultAddedCatalogType, [Validators.required]]
-    });
-  }
+  ) {}
 
   ngOnInit() {
-    this.store.state.clear();
-    this.typeCapabilities = Object.keys(TypeCapabilities);
+    if (this.data.addedCatalog) {
+      this.customData$.next(this.data.addedCatalog);
+    }
+    const types = Object.keys(TypeCapabilities).map((t) => {
+      return { value: t, title: t };
+    });
+    this.storeViewAll$$ = this.data.store.view.all$().subscribe(() => {
+      const predefinedChoices = this.data.predefinedCatalogs
+        .filter((c) => !this.data.store.get(c.id))
+        .map((predefinedCatalog) => {
+          return {
+            value: predefinedCatalog.id,
+            title: predefinedCatalog.title
+          };
+        });
 
-    this.addedCatalogType$$ = this.form
-      .get('type')
-      .valueChanges.subscribe((value) => {
-        if (value === 'wmts') {
-          this.form.get('title').setValidators(Validators.required);
-        } else {
-          this.form.get('title').setValidators([]);
+      const predefinedFieldConfigs = [
+        {
+          name: 'id',
+          title: this.languageService.translate.instant(
+            'igo.geo.catalog.library.predefined.select'
+          ),
+          type: 'select',
+          options: {
+            cols: 1,
+            validator: Validators.required
+          },
+          inputs: {
+            choices: predefinedChoices
+          }
         }
-        this.form.get('title').updateValueAndValidity();
-      });
-
-    this.computePredefinedCatalogList();
-    this.storeViewAll$$ = this.store.view
-      .all$()
-      .subscribe(() => this.computePredefinedCatalogList());
-
+      ];
+      const predefinedFormFields = predefinedFieldConfigs.map((config) =>
+        this.formService.field(config)
+      );
+      this.predefinedForm$.next(
+        this.formService.form(predefinedFormFields, [])
+      );
+    });
+    const customFieldConfigs = [
+      {
+        name: 'title',
+        title: this.languageService.translate.instant(
+          'igo.geo.catalog.library.custom.title'
+        ),
+        type: 'text',
+        options: {
+          validator: Validators.required
+        }
+      },
+      {
+        name: 'url',
+        title: 'URL',
+        type: 'text',
+        options: {
+          validator: Validators.required
+        }
+      },
+      {
+        name: 'type',
+        title: 'Type',
+        type: 'select',
+        options: {
+          cols: 1,
+          validator: Validators.required
+        },
+        inputs: {
+          choices: types
+        }
+      }
+    ];
+    const customFieldFormFields = customFieldConfigs.map((config) =>
+      this.formService.field(config)
+    );
+    this.customForm$.next(this.formService.form(customFieldFormFields, []));
     this.emailAddress = this.configService.getConfig('emailAddress');
   }
 
+  onPredefinedSubmit(data: Catalog) {
+    const catalog = this.data.predefinedCatalogs.find((c) => c.id === data.id);
+    if (this.checkFieldsValidity(this.predefinedForm$)) {
+      this.dialogRef.close(catalog);
+    }
+  }
+  onCustomSubmit(catalog: Catalog) {
+    if (this.checkFieldsValidity(this.customForm$)) {
+      this.dialogRef.close(catalog);
+    }
+  }
   ngOnDestroy() {
-    this.addedCatalogType$$.unsubscribe();
-    this.storeViewAll$$.unsubscribe();
+    this.storeViewAll$$?.unsubscribe();
   }
 
-  changeUrlOrTitle(catalog: Catalog) {
-    this.form.patchValue(catalog);
-    this.error = false;
-    this.computePredefinedCatalogList();
-  }
-
-  computePredefinedCatalogList() {
-    this.predefinedCatalogsList$.next(
-      this.predefinedCatalogs.filter((c) => !this.store.get(c.id))
-    );
-  }
-
-  addCatalog(addedCatalog: Catalog) {
-    this.error = false;
-    this.dialogRef.close(addedCatalog);
-  }
-
-  cancel() {
-    this.error = false;
-    this.dialogRef.close();
+  checkFieldsValidity(form$: BehaviorSubject<Form>): boolean {
+    const fields = form$.getValue().fields;
+    return fields
+      .map((field) => {
+        if (field.control.invalid) {
+          field.control.markAsTouched();
+          field.control.updateValueAndValidity();
+        }
+        return field.control.valid;
+      })
+      .every((e) => e);
   }
 }
