@@ -30,7 +30,9 @@ interface DataIssueReporterData {
 @Component({
   selector: 'igo-issue-reporter-tool',
   templateUrl: './data-issue-reporter-tool.component.html',
-  styleUrls: ['./data-issue-reporter-tool.component.scss']
+  styleUrls: ['./data-issue-reporter-tool.component.scss'],
+  // eslint-disable-next-line @angular-eslint/prefer-standalone
+  standalone: false
 })
 export class DataIssueReporterToolComponent implements OnInit, OnDestroy {
   /**
