@@ -15,7 +15,6 @@ import { MatListModule } from '@angular/material/list';
 import { MatSelectModule } from '@angular/material/select';
 import { MatSlideToggleModule } from '@angular/material/slide-toggle';
 
-import { IgoIconComponent } from '@igo2/common/icon';
 import { DetailedContext } from '@igo2/context';
 import { IgoLanguageModule } from '@igo2/core/language';
 import { AnyLayer, IgoMap, Layer, VectorLayer, isLayerItem } from '@igo2/geo';
@@ -27,7 +26,6 @@ import { MapState } from '../../map.state';
   selector: 'igo-advanced-swipe',
   templateUrl: './advanced-swipe.component.html',
   styleUrls: ['./advanced-swipe.component.scss'],
-  standalone: true,
   imports: [
     NgIf,
     FormsModule,
@@ -40,8 +38,7 @@ import { MapState } from '../../map.state';
     MatSlideToggleModule,
     MatListModule,
     MatIconModule,
-    IgoLanguageModule,
-    IgoIconComponent
+    IgoLanguageModule
   ]
 })
 export class AdvancedSwipeComponent implements OnInit, OnDestroy {
