@@ -2,14 +2,10 @@ import { JsonPipe } from '@angular/common';
 import { Component, Inject } from '@angular/core';
 import { MatButtonModule } from '@angular/material/button';
 
-import { provideAuthUserMonitoring } from '@igo2/auth/monitoring';
 import {
   AnyMonitoringOptions,
-  MONITORING_OPTIONS,
-  provideMonitoring
+  MONITORING_OPTIONS
 } from '@igo2/core/monitoring';
-
-import { environment } from 'projects/demo/src/environments/environment';
 
 import { DocViewerComponent } from '../../components/doc-viewer/doc-viewer.component';
 import { ExampleViewerComponent } from '../../components/example/example-viewer/example-viewer.component';
@@ -18,12 +14,7 @@ import { ExampleViewerComponent } from '../../components/example/example-viewer/
   selector: 'app-monitoring',
   templateUrl: './monitoring.component.html',
   styleUrls: ['./monitoring.component.scss'],
-  providers: [
-    JsonPipe,
-    ...provideMonitoring(environment.igo.monitoring),
-    ...provideAuthUserMonitoring(environment.igo.monitoring)
-  ],
-  standalone: true,
+  providers: [JsonPipe],
   imports: [DocViewerComponent, ExampleViewerComponent, MatButtonModule]
 })
 export class AppMonitoringComponent {
@@ -52,7 +43,7 @@ const EXAMPLE_PROVIDER = `bootstrapApplication(AppComponent, {
   providers: [
     ...provideMonitoring(environment.igo.monitoring),
 
-    // Provide the authentication user monitoring if you want to identify your user (id, fullname, email) in the error logging 
+    // Provide the authentication user monitoring if you want to identify your user (id, fullname, email) in the error logging
     ...provideAuthUserMonitoring(environment.igo.monitoring)
   ]
 })
